+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Additions
 - Added `BlockDatabase::size` and `StreamingDatabase::size`, which return the size in bytes of the database
 - Added `StreamingDatabase::stream_size`, which returns the size in bytes of a stream for the database
+- A new `asan` feature enables Address Sanitizer in the vendored version of `vectorscan`
 
 ## [v0.0.4](https://github.com/bradlarsen/vectorscan-rs/releases/v0.0.4) (2024-11-07)