+++ CHANGELOG.md
@@ -6,12 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Unreleased
+## [v0.0.5](https://github.com/bradlarsen/vectorscan-rs/releases/v0.0.5) (2024-12-09)
 
 ### Additions
 - Added `BlockDatabase::size` and `StreamingDatabase::size`, which return the size in bytes of the database
 - Added `StreamingDatabase::stream_size`, which returns the size in bytes of a stream for the database
 - A new `asan` feature enables Address Sanitizer in the vendored version of `vectorscan`
+
 
 ## [v0.0.4](https://github.com/bradlarsen/vectorscan-rs/releases/v0.0.4) (2024-11-07)